@@ -4,7 +4,7 @@ version = "0.0.0"
 authors = ["Near Inc <hello@nearprotocol.com>"]
 publish = false
 # Please update rust-toolchain.toml as well when changing version here:
-rust-version = "1.61.0"
+rust-version = "1.62.0"
 edition = "2021"
 
 [dependencies]
@@ -17,8 +17,7 @@ tracing = "0.1.13"
 itertools = "0.10.0"
 rand = "0.7"
 serde_json = "1"
-# Temporary workaround, fix with rust toolchain update.
-sysinfo = { git = "https://github.com/near/sysinfo", rev = "3cb97ee79a02754407d2f0f63628f247d7c65e7b" }
+sysinfo = "0.24.5"
 strum = { version = "0.24", features = ["derive"] }
 lru = "0.7.2"
 once_cell = "1.5.2"
