@@ -4,7 +4,7 @@ version = "0.0.0"
 authors = ["Near Inc <hello@nearprotocol.com>"]
 publish = false
 # Please update rust-toolchain.toml as well when changing version here:
-rust-version = "1.61.0"
+rust-version = "1.62.0"
 edition = "2021"
 
 [dependencies]
@@ -26,3 +26,6 @@ near-indexer-primitives = { path = "../indexer-primitives" }
 near-primitives = { path = "../../core/primitives" }
 near-store = { path = "../../core/store" }
 node-runtime = { path = "../../runtime/runtime" }
+
+[features]
+shardnet = ["nearcore/shardnet", "near-client/shardnet", "near-primitives/shardnet"]
