@@ -12,7 +12,6 @@ use crate::routing::edge_validator_actor::EdgeValidatorHelper;
 use crate::routing::routing_table_view::RoutingTableView;
 use crate::sink::Sink;
 use crate::stats::metrics;
-use crate::stats::metrics::{NetworkMetrics, PARTIAL_ENCODED_CHUNK_REQUEST_DELAY};
 use crate::store;
 use crate::types::{
     FullPeerInfo, NetworkClientMessages, NetworkInfo, NetworkRequests, NetworkResponses,
@@ -53,7 +52,7 @@ use std::sync::atomic::{AtomicUsize, Ordering};
 use std::sync::Arc;
 use tokio::net::{TcpListener, TcpStream};
 use tokio_stream::StreamExt;
-use tracing::{debug, error, info, trace, warn, Span};
+use tracing::{debug, error, info, trace, warn, Instrument, Span};
 use tracing_opentelemetry::OpenTelemetrySpanExt;
 
 /// How often to request peers from active peers.
@@ -179,8 +178,6 @@ pub struct PeerManagerActor {
     started_connect_attempts: bool,
     /// Connected peers we have sent new edge update, but we haven't received response so far.
     local_peer_pending_update_nonce_request: HashMap<PeerId, u64>,
-    /// Dynamic Prometheus metrics
-    pub(crate) network_metrics: NetworkMetrics,
     /// RoutingTableActor, responsible for computing routing table, routing table exchange, etc.
     routing_table_addr: Addr<routing::Actor>,
     /// Shared counter across all PeerActors, which counts number of `RoutedMessageBody::ForwardTx`
@@ -200,7 +197,7 @@ pub struct PeerManagerActor {
 pub enum Event {
     ServerStarted,
     RoutedMessageDropped,
-    RoutingTableUpdate(Arc<routing::RoutingTable>),
+    RoutingTableUpdate(Arc<routing::NextHopTable>),
     Ping(Ping),
     Pong(Pong),
 }
@@ -290,14 +287,18 @@ impl PeerManagerActor {
         let peer_store =
             PeerStore::new(&clock, store.clone(), &config.boot_nodes, config.blacklist.clone())
                 .map_err(|e| anyhow::Error::msg(e.to_string()))?;
-        debug!(target: "network", len = peer_store.len(), boot_nodes = config.boot_nodes.len(), "Found known peers");
+        debug!(target: "network",
+               len = peer_store.len(),
+               boot_nodes = config.boot_nodes.len(),
+               banned = peer_store.count_banned(),
+               "Found known peers");
         debug!(target: "network", blacklist = ?config.blacklist, "Blacklist");
 
         let my_peer_id = config.node_id();
         let network_graph = Arc::new(RwLock::new(routing::GraphWithCache::new(my_peer_id.clone())));
         let routing_table_addr =
             routing::Actor::new(clock.clone(), store.clone(), network_graph.clone()).start();
-        let routing_table_view = RoutingTableView::new(store);
+        let routing_table_view = RoutingTableView::new(store, my_peer_id.clone());
 
         let txns_since_last_block = Arc::new(AtomicUsize::new(0));
 
@@ -323,7 +324,6 @@ impl PeerManagerActor {
             routing_table_exchange_helper: Default::default(),
             started_connect_attempts: false,
             local_peer_pending_update_nonce_request: HashMap::new(),
-            network_metrics: Default::default(),
             routing_table_addr,
             txns_since_last_block,
             peer_counter: Arc::new(AtomicUsize::new(0)),
@@ -349,15 +349,17 @@ impl PeerManagerActor {
             .map(|response, act, _ctx| match response {
                 Ok(routing::actor::Response::RoutingTableUpdateResponse {
                     local_edges_to_remove,
-                    routing_table,
+                    next_hops,
                     peers_to_ban,
                 }) => {
-                    act.routing_table_view.remove_local_edges(local_edges_to_remove.iter());
-                    act.routing_table_view.peer_forwarding = routing_table.clone();
+                    for peer_id in &local_edges_to_remove {
+                        act.routing_table_view.remove_local_edge(peer_id);
+                    }
+                    act.routing_table_view.set_next_hops(next_hops.clone());
                     for peer in peers_to_ban {
                         act.ban_peer(&peer, ReasonForBan::InvalidEdge);
                     }
-                    act.event_sink.push(Event::RoutingTableUpdate(routing_table));
+                    act.event_sink.push(Event::RoutingTableUpdate(next_hops));
                 }
                 _ => error!(target: "network", "expected RoutingTableUpdateResponse"),
             })
@@ -368,24 +370,11 @@ impl PeerManagerActor {
         if edges.is_empty() {
             return;
         }
-        Self::add_local_edges(&mut self.routing_table_view, &edges, &self.my_peer_id);
 
-        self.routing_table_addr.do_send(routing::actor::Message::AddVerifiedEdges { edges });
-    }
-
-    fn add_local_edges(
-        routing_table_view: &mut RoutingTableView,
-        edges: &Vec<Edge>,
-        my_peer_id: &PeerId,
-    ) {
-        for edge in edges.iter() {
-            if let Some(other_peer) = edge.other(my_peer_id) {
-                if !routing_table_view.is_local_edge_newer(other_peer, edge.nonce()) {
-                    continue;
-                }
-                routing_table_view.local_edges_info.insert(other_peer.clone(), edge.clone());
-            }
+        for edge in &edges {
+            self.routing_table_view.add_local_edge(edge.clone());
         }
+        self.routing_table_addr.do_send(routing::actor::Message::AddVerifiedEdges { edges });
     }
 
     fn broadcast_accounts(&mut self, mut accounts: Vec<AnnounceAccount>) {
@@ -400,7 +389,6 @@ impl PeerManagerActor {
         }
 
         Self::broadcast_message(
-            self.network_metrics.clone(),
             &self.connected_peers,
             SendMessage {
                 message: PeerMessage::SyncRoutingTable(RoutingTableUpdate::from_accounts(accounts)),
@@ -471,7 +459,7 @@ impl PeerManagerActor {
         ctx: &mut Context<Self>,
         interval: time::Duration,
     ) {
-        let span =
+        let _span =
             tracing::trace_span!(target: "network", "broadcast_validated_edges_trigger").entered();
         let start = self.clock.now();
         let mut new_edges = Vec::new();
@@ -504,22 +492,18 @@ impl PeerManagerActor {
                         // We are not connected to this peer, but routing table contains
                         // information that we do. We should wait and remove that peer
                         // from routing table
-                        Self::wait_peer_or_remove(ctx, edge.clone(), self.network_metrics.clone());
+                        Self::wait_peer_or_remove(ctx, edge.clone());
                     }
-                    self.routing_table_view
-                        .local_edges_info
-                        .insert(other_peer.clone(), edge.clone());
+                    self.routing_table_view.add_local_edge(edge.clone());
                 }
             }
-            let network_metrics = self.network_metrics.clone();
-
             self.routing_table_addr
                 .send(routing::actor::Message::AddVerifiedEdges { edges: new_edges })
+                .in_current_span()
                 .into_actor(self)
                 .map(move |response, act, _ctx| {
                     let _span = tracing::trace_span!(
                         target: "network",
-                        parent: &span,
                         "broadcast_validated_edges_trigger_response")
                     .entered();
 
@@ -527,7 +511,6 @@ impl PeerManagerActor {
                         Ok(routing::actor::Response::AddVerifiedEdgesResponse(filtered_edges)) => {
                             // Broadcast new edges to all other peers.
                             Self::broadcast_message(
-                                network_metrics,
                                 &act.connected_peers,
                                 SendMessage {
                                     message: PeerMessage::SyncRoutingTable(
@@ -568,7 +551,7 @@ impl PeerManagerActor {
         throttle_controller: ThrottleController,
         ctx: &mut Context<Self>,
     ) {
-        let span = tracing::trace_span!(target: "network", "register_peer").entered();
+        let _span = tracing::trace_span!(target: "network", "register_peer").entered();
         debug!(target: "network", ?full_peer_info, "Consolidated connection");
 
         if self.outgoing_peers.contains(&full_peer_info.peer_info.id) {
@@ -607,31 +590,20 @@ impl PeerManagerActor {
 
         self.add_verified_edges_to_routing_table(vec![new_edge.clone()]);
 
-        let network_metrics = self.network_metrics.clone();
+        let run_later_span = tracing::trace_span!(target: "network", "RequestRoutingTableResponse");
         near_performance_metrics::actix::run_later(
             ctx,
             WAIT_FOR_SYNC_DELAY.try_into().unwrap(),
             move |act, ctx| {
-                let _span =
-                        tracing::trace_span!(target: "network", parent: &span, "RequestRoutingTableResponse")
-                            .entered();
+                let _guard = run_later_span.enter();
                 let known_edges = act.network_graph.read().edges().values().cloned().collect();
-                act.send_sync(
-                    network_metrics,
-                    peer_type,
-                    addr,
-                    ctx,
-                    target_peer_id.clone(),
-                    new_edge,
-                    known_edges,
-                );
+                act.send_sync(peer_type, addr, ctx, target_peer_id.clone(), new_edge, known_edges);
             },
         );
     }
 
     fn send_sync(
         &self,
-        network_metrics: NetworkMetrics,
         peer_type: PeerType,
         addr: Addr<PeerActor>,
         ctx: &mut Context<Self>,
@@ -639,14 +611,12 @@ impl PeerManagerActor {
         new_edge: Edge,
         known_edges: Vec<Edge>,
     ) {
-        let span = tracing::trace_span!(target: "network", "send_sync").entered();
+        let run_later_span = tracing::trace_span!(target: "network", "send_sync_attempt");
         near_performance_metrics::actix::run_later(
             ctx,
             WAIT_FOR_SYNC_DELAY.try_into().unwrap(),
             move |act, _ctx| {
-                let _span =
-                    tracing::trace_span!(target: "network", parent: &span, "send_sync_attempt")
-                        .entered();
+                let _guard = run_later_span.enter();
                 // Start syncing network point of view. Wait until both parties are connected before start
                 // sending messages.
                 let known_accounts = act.routing_table_view.get_announce_accounts();
@@ -671,7 +641,6 @@ impl PeerManagerActor {
                     // Only broadcast new message from the outbound endpoint.
                     // Wait a time out before broadcasting this new edge to let the other party finish handshake.
                     Self::broadcast_message(
-                        network_metrics,
                         &act.connected_peers,
                         SendMessage {
                             message: PeerMessage::SyncRoutingTable(RoutingTableUpdate::from_edges(
@@ -707,7 +676,6 @@ impl PeerManagerActor {
                 let edge_update = edge.remove_edge(self.my_peer_id.clone(), &self.config.node_key);
                 self.add_verified_edges_to_routing_table(vec![edge_update.clone()]);
                 Self::broadcast_message(
-                    self.network_metrics.clone(),
                     &self.connected_peers,
                     SendMessage {
                         message: PeerMessage::SyncRoutingTable(RoutingTableUpdate::from_edges(
@@ -935,7 +903,7 @@ impl PeerManagerActor {
         });
     }
 
-    fn wait_peer_or_remove(ctx: &mut Context<Self>, edge: Edge, network_metrics: NetworkMetrics) {
+    fn wait_peer_or_remove(ctx: &mut Context<Self>, edge: Edge) {
         // This edge says this is an connected peer, which is currently not in the set of connected peers.
         // Wait for some time to let the connection begin or broadcast edge removal instead.
 
@@ -948,7 +916,6 @@ impl PeerManagerActor {
                     // Peer is still not connected after waiting a timeout.
                     let new_edge = edge.remove_edge(act.my_peer_id.clone(), &act.config.node_key);
                     Self::broadcast_message(
-                        network_metrics,
                         &act.connected_peers,
                         SendMessage {
                             message: PeerMessage::SyncRoutingTable(RoutingTableUpdate::from_edges(
@@ -1251,12 +1218,8 @@ impl PeerManagerActor {
     }
 
     /// Broadcast message to all active peers.
-    fn broadcast_message(
-        network_metrics: NetworkMetrics,
-        connected_peers: &HashMap<PeerId, ConnectedPeer>,
-        msg: SendMessage,
-    ) {
-        network_metrics.inc_broadcast(msg.message.msg_variant());
+    fn broadcast_message(connected_peers: &HashMap<PeerId, ConnectedPeer>, msg: SendMessage) {
+        metrics::BROADCAST_MESSAGES.with_label_values(&[msg.message.msg_variant()]).inc();
         // TODO(MarX, #1363): Implement smart broadcasting. (MST)
 
         // Change message to reference counted to allow sharing with all actors
@@ -1348,7 +1311,7 @@ impl PeerManagerActor {
                       account_id = ?self.config.validator.as_ref().map(|v|v.account_id()),
                       to = ?msg.msg.target,
                       reason = ?find_route_error,
-                      known_peers = ?self.routing_table_view.peer_forwarding.len(),
+                      known_peers = ?self.routing_table_view.reachable_peers(),
                       msg = ?msg.msg.body,
                     "Drop signed message"
                 );
@@ -1495,7 +1458,6 @@ impl PeerManagerActor {
         match msg {
             NetworkRequests::Block { block } => {
                 Self::broadcast_message(
-                    self.network_metrics.clone(),
                     &self.connected_peers,
                     SendMessage {
                         message: PeerMessage::Block(block),
@@ -1600,7 +1562,7 @@ impl PeerManagerActor {
                 NetworkResponses::NoResponse
             }
             NetworkRequests::PartialEncodedChunkRequest { target, request, create_time } => {
-                PARTIAL_ENCODED_CHUNK_REQUEST_DELAY
+                metrics::PARTIAL_ENCODED_CHUNK_REQUEST_DELAY
                     .observe((self.clock.now() - create_time.0).as_seconds_f64());
                 let mut success = false;
 
@@ -1725,7 +1687,6 @@ impl PeerManagerActor {
             NetworkRequests::Challenge(challenge) => {
                 // TODO(illia): smarter routing?
                 Self::broadcast_message(
-                    self.network_metrics.clone(),
                     &self.connected_peers,
                     SendMessage {
                         message: PeerMessage::Challenge(challenge),
