@@ -48,7 +48,7 @@ impl Store {
     ) -> Result<(), Error> {
         let mut update = self.0.new_update();
         update.set::<schema::AccountAnnouncements>(account_id, aa);
-        update.commit().map_err(Error)
+        self.0.commit(update).map_err(Error)
     }
 
     /// Fetches row with key account_id from the AccountAnnouncements column.
@@ -79,7 +79,7 @@ impl Store {
         for peer_id in peers {
             update.set::<schema::PeerComponent>(peer_id, &component);
         }
-        update.commit().map_err(Error)
+        self.0.commit(update).map_err(Error)
     }
 
     /// Reads and deletes from DB the component that <peer_id> is a member of.
@@ -111,7 +111,7 @@ impl Store {
                 }
             }
         }
-        update.commit().map_err(Error)?;
+        self.0.commit(update).map_err(Error)?;
         Ok(edges)
     }
 }
@@ -126,14 +126,14 @@ impl Store {
     ) -> Result<(), Error> {
         let mut update = self.0.new_update();
         update.set::<schema::Peers>(peer_id, peer_state);
-        update.commit().map_err(Error)
+        self.0.commit(update).map_err(Error)
     }
 
     /// Deletes rows with keys in <peers> from Peers column.
     pub fn delete_peer_states(&mut self, peers: &[PeerId]) -> Result<(), Error> {
         let mut update = self.0.new_update();
         peers.iter().for_each(|p| update.delete::<schema::Peers>(p));
-        update.commit().map_err(Error)
+        self.0.commit(update).map_err(Error)
     }
 
     /// Reads the whole Peers column.
@@ -144,12 +144,12 @@ impl Store {
 
 impl From<near_store::Store> for Store {
     fn from(store: near_store::Store) -> Self {
-        Self(schema::Store::new(store))
+        Self(schema::Store::new(store.into_inner()))
     }
 }
 
 impl From<&near_store::Store> for Store {
     fn from(store: &near_store::Store) -> Self {
-        Self(schema::Store::new(store.clone()))
+        Self::from(store.clone())
     }
 }
