@@ -34,22 +34,21 @@ pub(crate) fn get_protocol_version_internal(
     client_protocol_version: ProtocolVersion,
     // Map of protocol versions to points in time when voting for that protocol version is expected to start.
     // If None or missing, the client votes for the latest protocol version immediately.
-    schedule: &HashMap<ProtocolVersion, Option<ProtocolUpgradeVotingSchedule>>,
+    schedule: &HashMap<ProtocolVersion, ProtocolUpgradeVotingSchedule>,
 ) -> ProtocolVersion {
     if next_epoch_protocol_version >= client_protocol_version {
-        return client_protocol_version;
-    }
-    match schedule.get(&client_protocol_version) {
-        Some(Some(voting_start)) => {
-            if voting_start.is_in_future() {
-                // Don't announce support for the latest protocol version yet.
-                next_epoch_protocol_version
-            } else {
-                // The time has passed, announce the latest supported protocol version.
-                client_protocol_version
-            }
+        client_protocol_version
+    } else if let Some(voting_start) = schedule.get(&client_protocol_version) {
+        if voting_start.is_in_future() {
+            // Don't announce support for the latest protocol version yet.
+            next_epoch_protocol_version
+        } else {
+            // The time has passed, announce the latest supported protocol version.
+            client_protocol_version
         }
-        _ => client_protocol_version,
+    } else {
+        // No schedule for latest supported version; go ahead and announce it.
+        client_protocol_version
     }
 }
 
@@ -98,8 +97,7 @@ mod tests {
         // As no protocol upgrade voting schedule is set, always return the version supported by the client.
 
         let client_protocol_version = 100;
-        let mut schedule = HashMap::new();
-        schedule.insert(client_protocol_version, None);
+        let schedule = HashMap::new();
 
         assert_eq!(
             client_protocol_version,
@@ -133,7 +131,7 @@ mod tests {
         let mut schedule = HashMap::new();
         schedule.insert(
             client_protocol_version,
-            Some(ProtocolUpgradeVotingSchedule::from_str("2050-01-01 00:00:00").unwrap()),
+            ProtocolUpgradeVotingSchedule::from_str("2050-01-01 00:00:00").unwrap(),
         );
 
         // The client supports a newer version than the version of the next epoch.
@@ -177,7 +175,7 @@ mod tests {
         let mut schedule = HashMap::new();
         schedule.insert(
             client_protocol_version,
-            Some(ProtocolUpgradeVotingSchedule::from_str("1900-01-01 00:00:00").unwrap()),
+            ProtocolUpgradeVotingSchedule::from_str("1900-01-01 00:00:00").unwrap(),
         );
 
         // Regardless of the protocol version of the next epoch, return the version supported by the client.
