@@ -1,7 +1,10 @@
-use crate::types::Balance;
+use std::collections::HashMap;
+use std::str::FromStr;
+
 use once_cell::sync::Lazy;
 use serde::{Deserialize, Serialize};
-use std::collections::HashMap;
+
+use crate::types::Balance;
 
 /// Data structure for semver version and github tag or commit.
 #[cfg_attr(feature = "deepsize_feature", derive(deepsize::DeepSizeOf))]
@@ -162,6 +165,9 @@ pub enum ProtocolFeature {
     /// Charge for contract loading before it happens.
     #[cfg(feature = "protocol_feature_fix_contract_loading_cost")]
     FixContractLoadingCost,
+    /// Charge for contract loading before it happens.
+    #[cfg(feature = "protocol_feature_account_id_in_function_call_permission")]
+    AccountIdInFunctionCallPermission,
 }
 
 /// Both, outgoing and incoming tcp connections to peers, will be rejected if `peer's`
@@ -176,7 +182,7 @@ const STABLE_PROTOCOL_VERSION: ProtocolVersion = 55;
 cfg_if::cfg_if! {
     if #[cfg(feature = "nightly_protocol")] {
         /// Current latest nightly version of the protocol.
-        pub const PROTOCOL_VERSION: ProtocolVersion = 129;
+        pub const PROTOCOL_VERSION: ProtocolVersion = 130;
     } else if #[cfg(feature = "shardnet")] {
         /// Protocol version for shardnet.
         pub const PROTOCOL_VERSION: ProtocolVersion = 100;
@@ -187,13 +193,14 @@ cfg_if::cfg_if! {
 
 /// The points in time after which the voting for the protocol version should start.
 #[allow(dead_code)]
-const PROTOCOL_UPGRADE_SCHEDULE: Lazy<
-    HashMap<ProtocolVersion, Option<ProtocolUpgradeVotingSchedule>>,
-> = Lazy::new(|| {
-    let mut schedule = HashMap::new();
-    schedule.insert(52, None);
-    schedule
-});
+const PROTOCOL_UPGRADE_SCHEDULE: Lazy<HashMap<ProtocolVersion, ProtocolUpgradeVotingSchedule>> =
+    Lazy::new(|| {
+        let mut schedule = HashMap::new();
+        // Update to latest protocol version on release.
+        schedule
+            .insert(54, ProtocolUpgradeVotingSchedule::from_str("2022-06-27 15:00:00").unwrap());
+        schedule
+    });
 
 /// Gives new clients an option to upgrade without announcing that they support the new version.
 /// This gives non-validator nodes time to upgrade. See https://github.com/near/NEPs/issues/205
@@ -201,7 +208,7 @@ pub fn get_protocol_version(next_epoch_protocol_version: ProtocolVersion) -> Pro
     get_protocol_version_internal(
         next_epoch_protocol_version,
         PROTOCOL_VERSION,
-        &*PROTOCOL_UPGRADE_SCHEDULE,
+        &PROTOCOL_UPGRADE_SCHEDULE,
     )
 }
 
@@ -247,6 +254,8 @@ impl ProtocolFeature {
             ProtocolFeature::FixStakingThreshold => 126,
             #[cfg(feature = "protocol_feature_fix_contract_loading_cost")]
             ProtocolFeature::FixContractLoadingCost => 129,
+            #[cfg(feature = "protocol_feature_account_id_in_function_call_permission")]
+            ProtocolFeature::AccountIdInFunctionCallPermission => 130,
         }
     }
 }
