@@ -1,5 +1,6 @@
 use std::collections::{HashMap, HashSet};
 
+use genesis_populate::get_account_id;
 use near_crypto::{InMemorySigner, KeyType};
 use near_primitives::hash::CryptoHash;
 use near_primitives::transaction::{Action, FunctionCallAction, SignedTransaction};
@@ -7,7 +8,6 @@ use near_primitives::types::AccountId;
 use rand::prelude::ThreadRng;
 use rand::Rng;
 
-use crate::utils::get_account_id;
 /// A helper to create transaction for processing by a `TestBed`.
 #[derive(Clone)]
 pub(crate) struct TransactionBuilder {
@@ -87,7 +87,7 @@ impl TransactionBuilder {
         rand::thread_rng()
     }
 
-    pub(crate) fn account(&mut self, account_index: usize) -> AccountId {
+    pub(crate) fn account(&mut self, account_index: u64) -> AccountId {
         get_account_id(account_index)
     }
     pub(crate) fn random_account(&mut self) -> AccountId {
