@@ -155,9 +155,12 @@ impl DumpStateRedisCmd {
 
 #[derive(Parser)]
 pub struct DumpTxCmd {
-    /// Specify which block to dump transactions for.
+    /// Specify the start block by height to begin dumping transactions from, inclusive.
     #[clap(long)]
-    height: BlockHeight,
+    start_height: BlockHeight,
+    /// Specify the end block by height to stop dumping transactions at, inclusive.
+    #[clap(long)]
+    end_height: BlockHeight,
     /// List of account IDs to dump.
     /// If not set, all account IDs will be dumped.
     #[clap(long)]
@@ -170,7 +173,8 @@ pub struct DumpTxCmd {
 impl DumpTxCmd {
     pub fn run(self, home_dir: &Path, near_config: NearConfig, store: Store) {
         dump_tx(
-            self.height,
+            self.start_height,
+            self.end_height,
             home_dir,
             near_config,
             store,
